+++ src/context/index.ts
@@ -1,2 +1,3 @@
 export * from "./AuthContext"
 export * from "./LayoutContext"
+export * from "./RightPanelContext"