+++ src/hooks/useRightPanel.ts
@@ -1,0 +1,7 @@
+import { useContext } from "react"
+
+import { RightPanelContext } from "context"
+
+export const useRightPanel = () => {
+  return useContext(RightPanelContext)
+}