+++ src/pages/create/user.tsx
@@ -7,7 +7,7 @@ import { request } from "graphql-request"
 
 import { SlackLayout } from "components/SlackLayout"
 
-const ENDPOINT = "http://localhost:4000"
+const ENDPOINT = process.env.ENDPOINT as string
 
 const FRAGMENT = `
 fragment Shared on User {